+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add support for saving Azure key version with DEK (#2641)
 * Pass context when clients make KEK calls to DEK Registry (#2642)
+* Add support for inline validation rules (#2651)
 
 ## Fixes